requirements.txt
influxdb
python-daemon
pyyaml